added receiveIteration code. Need to debug crash on event
