+++ tests/fm_dev/rti1516eLv/rti1516eLv.cpp
@@ -29,42 +29,9 @@ namespace rti1516eLv
     bool _disconnected = false;
     bool _processCB = false;
 
-    // temp LVuser event store
-    LVUserEventRef tempUserEvStore;
     LVUserEventRef lueObjInsNameResSucceeded;
     LVUserEventRef lueObjInsNameResFailed;
     LVUserEventRef lueReceiveInteraction;
-
-    wstring chararray2wstring(const char charArray[])
-    {
-        wstring wStringOut;
-        // allocate strings
-        string sStringTemp(charArray);
-        wStringOut.assign(sStringTemp.begin(),sStringTemp.end());
-
-        return wStringOut;
-    }
-
-    LStrHandle wstring2LvString(wstring const & theObjectInstanceName)
-    {
-            const wchar_t *input = theObjectInstanceName.c_str();
-            size_t size = (wcslen(input)+1)*sizeof(wchar_t);
-            char *buffer = new char[size];
-            wcstombs(buffer,input,size);
-            // char buffer[] = "Ciao Pirla";
-            const int headerSize = 4;
-            size = strlen(buffer);
-
-	        // Allocate memory for the LV string
-            LStrHandle lSh = (LStrHandle)DSNewHandle(headerSize+size);
-
-            // Write the LV string header
-            (*lSh)->cnt = size;
-            memcpy( (*lSh)->str,buffer,size);
-            
-
-        return lSh;
-    }
 
     class LvFederate : public NullFederateAmbassador {
     private:
@@ -403,10 +370,15 @@ namespace rti1516eLv
             throw (FederateInternalError)
         {
             receiveInteractionData dataToSend;
+
+            dataToSend.interactionClassHandle = theInteraction;
+            dataToSend.parHandleValueMap = theParameterValues;
             dataToSend.sentOrder = sentOrder;
             dataToSend.theType = theType;
-            PostLVUserEvent(lueReceiveInteraction,&dataToSend);
             
+            
+            PostLVUserEvent(lueReceiveInteraction,&dataToSend);
+
             if (theInteraction == _iMessageId) {
                 HLAunicodeString message;
                 HLAunicodeString sender;
@@ -487,104 +459,6 @@ namespace rti1516eLv
     };
     
     LvFederate* oLvFederate; 
-
-    // test export functions
-
-    EXTERNC int testFunc()
-    {
-        return 12345;
-    }
-
-    // using struct to test complex data sharing
-    EXTERNC MgErr testFireEvent(testEventData *value)
-    {
-        MgErr status;
-        status = PostLVUserEvent(tempUserEvStore,value);
-        return status;
-    }
-
-    EXTERNC int regObjInstNameResSuccEvent(LVUserEventRef *eventRef)
-    {
-        tempUserEvStore = *eventRef;
-        return 0;
-    }
-
-    // helpers 
-
-    EXTERNC  int attrHandleValueMapCreate(AttributeHandleValueMap **attrHandleValueMap)
-    {
-        *attrHandleValueMap = new AttributeHandleValueMap();
-        return 0;
-    }
-
-    EXTERNC int attrHandleValueMapAddElementString(
-        AttributeHandleValueMap * attrHandleValueMap,
-        AttributeHandle *attributeHandle,
-        const char sElem[])
-    {
-        wstring wElem = chararray2wstring(sElem);
-        HLAunicodeString uElem(wElem);
-        (*attrHandleValueMap)[(*attributeHandle)] = uElem.encode();
-        
-    }
-
-    EXTERNC int attrHandleValueMapDestroy(
-        AttributeHandleValueMap * attrHandleValueMap)
-    {
-        delete attrHandleValueMap;
-
-        return 0;
-    }
-
-    EXTERNC int parHandleValueMapCreate(ParameterHandleValueMap **parHandleValueMap)
-    {
-        *parHandleValueMap = new ParameterHandleValueMap();
-        return 0;
-    }
-
-    EXTERNC int parHandleValueMapAddElementString(
-        ParameterHandleValueMap * parHandleValueMap,
-        ParameterHandle * parameterHandle,
-        const char sElem[])
-    {
-        wstring wElem = chararray2wstring(sElem);
-        HLAunicodeString uElem(wElem);
-        (*parHandleValueMap)[(*parameterHandle)] = uElem.encode();
-    }
-
-    EXTERNC int parHandleValueMapDestroy(
-        ParameterHandleValueMap * parHandleValueMap)
-    {
-        delete parHandleValueMap;
-
-        return 0;
-    } 
-
-    EXTERNC int attrHandleSetCreate(AttributeHandleSet **attrHandleSet)
-    {
-        *attrHandleSet = new AttributeHandleSet();
-        return 0;
-    }
-
-    EXTERNC int attrHandleSetInsert(
-        AttributeHandleSet *attrHandleSet,
-        AttributeHandle *attrHandle
-    )
-    {
-        //AttributeHandle attrHandleVar;
-        //attrHandleSet->insert(attrHandleVar);
-        //attrHandleSet->insert(*attrHandle);
-        (*attrHandleSet).insert(*attrHandle);
-        return 0;
-    }
-
-    EXTERNC int attrHandleSetDestroy(
-        AttributeHandleSet * attrHandleSet
-    )
-    {
-        delete attrHandleSet;
-        return 0;
-    }
 
     // OpenRTI wrappers
