+++ tests/fm_dev/rti1516eLv/rti1516eLv.h
@@ -27,28 +27,14 @@ using namespace rti1516e;
 
 namespace rti1516eLv
 {
-    // data types
-    struct testEventData {
-        int count1;
-        int count2;
-    };
 
     struct receiveInteractionData {
+        InteractionClassHandle interactionClassHandle;
+        ParameterHandleValueMap parHandleValueMap;
         OrderType sentOrder;
         TransportationType theType;
     };
     
-    // internal functions
-    int lvErrorCodeFromException(Exception &e);
-    //wstring chararray2wstring(const char charArray[]);
-
-    // Test or debug functions
-    EXTERNC int testFunc();
-    EXTERNC MgErr testFireEvent(
-        testEventData *value);
-    EXTERNC int regObjInstNameResSuccEvent(
-        LVUserEventRef *eventRef);
-
     EXTERNC int regLvUserEvents(
         LVUserEventRef *objInstNameResSucc,
         LVUserEventRef *objInstNameResFail,
