+++ tests/fm_dev/rti1516eLv/rti1516eLv_errors.h
@@ -149,6 +149,7 @@
 #define	LV_ERROR_INTERNALERROR	LV_ERROR_RANGE_OFFSET-121
 
 //Ambassador specific warning codes
+#define	LV_WARNING_FEDERATESCURRENTLYJOINED	LV_WARNING_RANGE_OFFSET+49
 #define	LV_WARNING_FEDERATIONEXECUTIONALREADYEXISTS LV_WARNING_RANGE_OFFSET+50
 
 //#define LV_ERROR_INVALID_LOCAL_SETTINGS_DESIGNATOR  LV_ERROR_RANGE_OFFSET-2
